converted from previously reading 16bit instructions to instead reading 2 8bit addresses, aligning with CHIP-8

+++ read_file/read_file.h
@@ -9,4 +9,4 @@ FILE*  get_program_file(char *program_file_name);
 int close_program_file(FILE* file);
 
 uint8_t* read_two_bytes(FILE* file, uint8_t* two_byte_buffer);
-uint16_t* load_program_file(char *program_file_name);
+uint8_t* load_program_file(char *program_file_name);
